refactor: Parser->StringParser 클래스명 변경 및 메서드명 변경
