+++ src/main/java/domain/StringParser.java
@@ -1,14 +1,14 @@
-package utils;
+package domain;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class Parser {
+public class StringParser {
 
     private static final int SPLIT_LIMIT = -1;
 
-    public static List<String> parse(String input, String delimiter) {
+    public static List<String> splitByComma(String input, String delimiter) {
         return Arrays.stream(input.split(delimiter, SPLIT_LIMIT))
                 .collect(Collectors.toList());
     }
